좋아요 수정

+++ src/main/java/com/proj/instagram/post/LikeDTO.java
@@ -3,17 +3,17 @@ package com.proj.instagram.post;
 import java.util.Date;
  
 public class LikeDTO {
-    private Long likeId;        // 좋아요 ID
+    private int likeId;        // 좋아요 ID
     private String accountId;   // 사용자 계정 ID (INSTA_USER의 EMAIL)
     private int postId;        // 게시물 ID (INSTA_POST의 POST_ID)
     private Date likedAt;       // 좋아요한 날짜
  
     // Getters and Setters
-    public Long getLikeId() {
+    public int getLikeId() {
         return likeId;
     }
 
-    public void setLikeId(Long likeId) { 
+    public void setLikeId(int likeId) { 
         this.likeId = likeId;
     }
